Finished UI, handling passkey interrupt, some clean up
