--- packages/app/features/auth/sign-up/NewSignUpForm.tsx
@@ -1,365 +1,0 @@
-import {
-  Button,
-  FadeCard,
-  Paragraph,
-  Separator,
-  SubmitButton,
-  useToastController,
-  XStack,
-  YStack,
-} from '@my/ui'
-import { useCallback, useEffect, useState } from 'react'
-import { Turnstile } from 'app/features/auth/sign-up/Turnstile'
-import { FormProvider, useForm } from 'react-hook-form'
-import { z } from 'zod'
-import { SendtagSchema } from 'app/utils/zod/sendtag'
-import { formFields, SchemaForm } from 'app/utils/SchemaForm'
-import { signChallenge } from 'app/utils/signChallenge'
-import { bytesToHex, hexToBytes } from 'viem'
-import { RecoveryOptions } from '@my/api/src/routers/account-recovery/types'
-import { formatErrorMessage } from 'app/utils/formatErrorMessage'
-import { api } from 'app/utils/api'
-import { useRouter } from 'solito/router'
-import { useAuthScreenParams } from 'app/routers/params'
-import { SendtagAvailability } from '@my/api/src/routers/tag/types'
-import { useSendAccount } from 'app/utils/send-accounts'
-import { useSupabase } from 'app/utils/supabase/useSupabase'
-import { base16, base64urlnopad } from '@scure/base'
-import { asciiToByteArray } from 'app/utils/asciiToByteArray'
-import { createPasskey } from 'app/utils/createPasskey'
-import { base64URLNoPadToBase16 } from 'app/utils/base64ToBase16'
-import { assert } from 'app/utils/assert'
-
-const SendtagSchemaWithoutRestrictions = z.object({
-  name: formFields.text,
-})
-
-export const NewSignUpForm = () => {
-  const [captchaToken, setCaptchaToken] = useState<string | undefined>()
-  const [isInputFocused, setIsInputFocused] = useState<boolean>(false)
-  const [isSigningIn, setIsSigningIn] = useState<boolean>(false)
-  const form = useForm<z.infer<typeof SendtagSchemaWithoutRestrictions>>()
-  const router = useRouter()
-  const [queryParams] = useAuthScreenParams()
-  const { redirectUri } = queryParams
-  const toast = useToastController()
-  const sendAccount = useSendAccount()
-  const supabase = useSupabase()
-
-  const formName = form.watch('name')
-  const validationError = form.formState.errors.root
-  const canSubmit = !!formName && !validationError
-
-  const { refetch: refetchCheckSendtagAvailability } = api.tag.checkAvailability.useQuery(
-    { name: formName },
-    { enabled: false }
-  )
-
-  const { mutateAsync: signUpMutateAsync } = api.auth.signUp.useMutation({
-    retry: false,
-  })
-
-  const { mutateAsync: getChallengeMutateAsync } = api.challenge.getChallenge.useMutation({
-    retry: false,
-  })
-
-  const { mutateAsync: validateSignatureMutateAsync } = api.challenge.validateSignature.useMutation(
-    { retry: false }
-  )
-
-  const { mutateAsync: sendAccountCreateMutateAsync } = api.sendAccount.create.useMutation()
-
-  useEffect(() => {
-    const subscription = form.watch(() => {
-      form.clearErrors('root')
-    })
-
-    return () => subscription.unsubscribe()
-  }, [form.watch, form.clearErrors])
-
-  const handleSubmit = async ({ name }: z.infer<typeof SendtagSchemaWithoutRestrictions>) => {
-    const { error: schemaError } = SendtagSchema.safeParse({ name })
-
-    if (schemaError) {
-      form.setError('root', {
-        type: 'custom',
-        message: schemaError.errors[0]?.message ?? 'Invalid Sendtag',
-      })
-
-      return
-    }
-
-    const { data: checkSendtagAvailabilityResponse, error: checkSendtagAvailabilityError } =
-      await refetchCheckSendtagAvailability()
-
-    if (checkSendtagAvailabilityError) {
-      form.setError('root', {
-        type: 'custom',
-        message: checkSendtagAvailabilityError.message ?? 'Error checking sendtag availability',
-      })
-
-      return
-    }
-
-    if (
-      checkSendtagAvailabilityResponse &&
-      checkSendtagAvailabilityResponse.sendtagAvailability === SendtagAvailability.Taken
-    ) {
-      form.setError('root', {
-        type: 'custom',
-        message: 'This Sendtag is already taken',
-      })
-
-      return
-    }
-
-    try {
-      await signUpMutateAsync({
-        sendtag: name,
-        captchaToken,
-      })
-
-      const { data: sessionData } = await supabase.auth.getSession()
-
-      assert(!!sessionData?.session?.user?.id, 'No user id')
-
-      const { data: sendAcct, error: refetchError } = await sendAccount.refetch()
-      if (refetchError) throw refetchError
-      if (sendAcct) {
-        throw new Error(`Account already created: ${sendAcct.address}`)
-      }
-
-      const accountName = name
-      const keySlot = 0
-      const passkeyName = `${sessionData.session.user.id}.${keySlot}` // 64 bytes max
-      const challenge = base64urlnopad.encode(asciiToByteArray('foobar'))
-
-      const [rawCred, authData] = await createPasskey({
-        user: sessionData.session.user,
-        keySlot,
-        challenge,
-        accountName,
-      })
-
-      const raw_credential_id = base64URLNoPadToBase16(rawCred.rawId)
-      const attestation_object = base64URLNoPadToBase16(rawCred.response.attestationObject)
-
-      await sendAccountCreateMutateAsync({
-        accountName,
-        passkeyName,
-        rawCredentialIDB16: raw_credential_id,
-        cosePublicKeyB16: base16.encode(authData.COSEPublicKey),
-        rawAttestationObjectB16: attestation_object,
-        keySlot,
-      })
-
-      const { data: sendAcct2, error: refetchError2 } = await sendAccount.refetch()
-      if (refetchError2) throw refetchError
-      if (sendAcct2) {
-        router.replace('/')
-        return
-      }
-
-      form.setError('root', {
-        type: 'custom',
-        message: 'Account not created. Please try again.',
-      })
-
-      // TODO register sendtag
-    } catch (error) {
-      form.setError('root', {
-        type: 'custom',
-        message: error.message,
-      })
-    }
-  }
-
-  const handleSignIn = useCallback(async () => {
-    setIsSigningIn(true)
-
-    try {
-      const challengeData = await getChallengeMutateAsync()
-
-      const rawIdsB64: { id: string; userHandle: string }[] = []
-      const { encodedWebAuthnSig, accountName, keySlot } = await signChallenge(
-        challengeData.challenge as `0x${string}`,
-        rawIdsB64
-      )
-
-      const encodedWebAuthnSigBytes = hexToBytes(encodedWebAuthnSig)
-      const newEncodedWebAuthnSigBytes = new Uint8Array(encodedWebAuthnSigBytes.length + 1)
-      newEncodedWebAuthnSigBytes[0] = keySlot
-      newEncodedWebAuthnSigBytes.set(encodedWebAuthnSigBytes, 1)
-
-      await validateSignatureMutateAsync({
-        recoveryType: RecoveryOptions.WEBAUTHN,
-        signature: bytesToHex(newEncodedWebAuthnSigBytes),
-        challengeId: challengeData.id,
-        identifier: `${accountName}.${keySlot}`,
-      })
-
-      router.push(redirectUri ?? '/')
-    } catch (error) {
-      toast.show(formatErrorMessage(error), {
-        preset: 'error',
-        isUrgent: true,
-        duration: 10000000,
-      })
-    } finally {
-      setIsSigningIn(false)
-    }
-  }, [getChallengeMutateAsync, router, redirectUri, validateSignatureMutateAsync, toast.show])
-
-  const renderAfterContent = useCallback(
-    ({ submit }: { submit: () => void }) => (
-      <SubmitButton
-        alignSelf={'center'}
-        w={'90%'}
-        theme="green"
-        onPress={submit}
-        py={'$5'}
-        br={'$4'}
-        variant={canSubmit ? undefined : 'outlined'}
-        bw={'$1'}
-        disabled={!canSubmit}
-      >
-        <Button.Text
-          ff={'$mono'}
-          fontWeight={'500'}
-          tt="uppercase"
-          size={'$5'}
-          color={canSubmit ? '$black' : '$primary'}
-        >
-          create account
-        </Button.Text>
-      </SubmitButton>
-    ),
-    [canSubmit]
-  )
-
-  return (
-    <YStack f={1} jc={'space-around'} ai={'center'} gap={'$3.5'} py={'$3.5'}>
-      <FormProvider {...form}>
-        <YStack w={'100%'} ai={'center'}>
-          <Paragraph w={'90%'} size={'$8'} fontWeight={500} tt={'uppercase'}>
-            create your account
-          </Paragraph>
-          <Paragraph w={'90%'} size={'$5'} color={'$olive'}>
-            Sendtags are usernames
-          </Paragraph>
-          <SchemaForm
-            form={form}
-            onSubmit={handleSubmit}
-            schema={SendtagSchemaWithoutRestrictions}
-            defaultValues={{
-              name: '',
-            }}
-            props={{
-              name: {
-                placeholder: 'Input desired Sendtag',
-                color: '$color12',
-                fontWeight: '500',
-                bw: 0,
-                br: 0,
-                p: 0,
-                pl: '$2.5',
-
-                focusStyle: {
-                  outlineWidth: 0,
-                },
-                '$theme-dark': {
-                  placeholderTextColor: '$darkGrayTextField',
-                },
-                '$theme-light': {
-                  placeholderTextColor: '$darkGrayTextField',
-                },
-                fontSize: '$5',
-                onFocus: () => setIsInputFocused(true),
-                onBlur: () => setIsInputFocused(false),
-                fieldsetProps: {
-                  width: '100%',
-                },
-                iconBefore: (
-                  <Paragraph ml={-12} size={'$5'} opacity={formName ? 1 : 0}>
-                    /
-                  </Paragraph>
-                ),
-              },
-            }}
-            formProps={{
-              w: '100%',
-              footerProps: { pb: 0 },
-              $gtSm: {
-                maxWidth: '100%',
-              },
-              style: { justifyContent: 'space-between' },
-            }}
-            renderAfter={renderAfterContent}
-          >
-            {({ name }) => {
-              return (
-                <FadeCard
-                  w={'100%'}
-                  pb={validationError ? '$5' : '$6'}
-                  my={'$5'}
-                  borderColor={validationError ? '$error' : 'transparent'}
-                  bw={1}
-                >
-                  <XStack position="relative">
-                    {name}
-                    <XStack
-                      position="absolute"
-                      bottom={0}
-                      left={0}
-                      right={0}
-                      height={1}
-                      backgroundColor={isInputFocused ? '$primary' : '$darkGrayTextField'}
-                      $theme-light={{
-                        backgroundColor: isInputFocused ? '$color12' : '$silverChalice',
-                      }}
-                    />
-                  </XStack>
-                  {validationError && (
-                    <Paragraph color={'$error'}>{validationError.message}</Paragraph>
-                  )}
-                </FadeCard>
-              )
-            }}
-          </SchemaForm>
-        </YStack>
-        <YStack w={'100%'} gap={'$3.5'}>
-          <XStack w={'100%'} gap={'$3.5'} ai={'center'}>
-            <Separator bc={'$color10'} />
-            <Paragraph tt={'uppercase'}>or</Paragraph>
-            <Separator bc={'$color10'} />
-          </XStack>
-          <Button
-            onPress={handleSignIn}
-            transparent
-            chromeless
-            backgroundColor="transparent"
-            hoverStyle={{ backgroundColor: 'transparent' }}
-            pressStyle={{ backgroundColor: 'transparent' }}
-            focusStyle={{ backgroundColor: 'transparent' }}
-            bw={0}
-            br={0}
-            height={'auto'}
-            p={'$3'}
-            disabled={isSigningIn}
-          >
-            <Button.Text
-              color={'$primary'}
-              $theme-light={{
-                color: '$color12',
-              }}
-            >
-              {isSigningIn ? 'Signing in...' : 'Sign in'}
-            </Button.Text>
-          </Button>
-        </YStack>
-        <Turnstile onSuccess={(t) => setCaptchaToken(t)} />
-        <Paragraph>TODO: create a backup</Paragraph>
-      </FormProvider>
-    </YStack>
-  )
-}
